cambios 2

+++ src/services/whatsappService.js
@@ -9,21 +9,29 @@ const fs = require("fs");
 const socketManager = require("../utils/socketManager");
 const log = (pino = require("pino"));
 
+const activeConnections = new Map();
+
 exports.generateUniqueDeviceID = () => {
   let newDeviceID = Math.floor(Math.random() * 10) + 1;
   return newDeviceID;
 };
 
-exports.connectOrReconnect = async (userID) => {
-  let codeQR;
-  const sessionDir = `session_${userID}`;
-  const { state, saveCreds } = await useMultiFileAuthState(sessionDir);
-  const sock = makeWASocket({
-    printQRInTerminal: false,
-    auth: state,
-    logger: log({ level: "silent" }),
-  });
-  return new Promise((resolve, reject) => {
+exports.connectOrReconnect = async (userID, message) => {
+  if (activeConnections.has(userID) && !!message) {
+    console.log(`Conexión ya existente para el usuario ${userID}`);
+    return activeConnections.get(userID);
+  }
+
+  let connectionPromise = new Promise(async (resolve, reject) => {
+    let codeQR;
+    const sessionDir = `session_${userID}`;
+    const { state, saveCreds } = await useMultiFileAuthState(sessionDir);
+    const sock = makeWASocket({
+      printQRInTerminal: false,
+      auth: state,
+      logger: log({ level: "silent" }),
+    });
+
     sock.ev.on("connection.update", async (update) => {
       const { connection, lastDisconnect, qr } = update;
       if (connection === "close") {
@@ -38,10 +46,18 @@ exports.connectOrReconnect = async (userID) => {
           );
           setTimeout(() => {
             console.log("Reiniciando proceso de conexión...");
-            this.connectOrReconnect(userID);
+            this.connectOrReconnect(userID, false);
           }, 5000);
         } else {
-          this.connectOrReconnect(userID);
+          this.connectOrReconnect(userID, false);
+        }
+        if (reason === DisconnectReason.connectionClosed) {
+          this.connectOrReconnect(userID, false);
+        } else if (reason === DisconnectReason.timedOut) {
+          reject({
+            message: "Request Time-out",
+            status: 408,
+          });
         }
       }
 
@@ -55,6 +71,16 @@ exports.connectOrReconnect = async (userID) => {
         });
       } else if (!qr && connection === "open") {
         socketManager.saveSockets();
+        activeConnections.set(
+          userID,
+          Promise.resolve({
+            message: "Conexión abierta",
+
+            status: 201,
+            deviceId: userID,
+            qr: null,
+          })
+        );
         resolve({
           message: "Conexión abierta",
           status: 201,
@@ -67,40 +93,47 @@ exports.connectOrReconnect = async (userID) => {
     socketManager.setSocket(userID, sock);
     sock.ev.on("creds.update", saveCreds);
   });
+
+  return connectionPromise;
 };
 
 exports.disconnectDevice = async (deviceId) => {
-  const connectionResult = await this.connectOrReconnect(deviceId);
+  if (activeConnections.has(deviceId)) {
+    const connectionResult = await activeConnections.get(deviceId);
+    if (connectionResult.status === 201) {
+      const sock = socketManager.getSocket(deviceId);
+      if (sock) {
+        try {
+          await sock.logout();
+          socketManager.removeSocket(deviceId);
+          const sessionDir = `session_${deviceId}`;
+          await fs.promises.rm(sessionDir, { recursive: true, force: true });
+          socketManager.saveSockets();
 
-  if (connectionResult.status === 201) {
-    return res.status(404).json({
-      error: `El ID del dispositivo no se encuentra registrado`,
-      status: 404,
-    });
-  }
+          activeConnections.delete(deviceId);
 
-  const sock = socketManager.getSocket(deviceId);
-  if (sock) {
-    try {
-      await sock.logout();
-      socketManager.removeSocket(deviceId);
-      const sessionDir = `session_${deviceId}`;
-      await fs.promises.rm(sessionDir, { recursive: true, force: true });
-      socketManager.saveSockets();
-
-      return {
-        message: `El usuario ha sido desconectado y la sesión eliminada`,
-        status: 200,
-      };
-    } catch (error) {
-      return {
-        message: `Error al desconectar el dispositivo ${deviceId}, usuario no encontrado o no esta vinculado`,
-        status: 404,
-      };
+          return {
+            message: `El usuario ha sido desconectado y la sesión eliminada`,
+            status: 200,
+          };
+        } catch (error) {
+          console.error(
+            `Error al desconectar el dispositivo ${deviceId}:`,
+            error
+          );
+          return {
+            message: `Error al desconectar el dispositivo ${deviceId}`,
+            status: 500,
+          };
+        }
+      }
     }
-  } else {
-    throw new Error(`Dispositivo ${deviceId} no encontrado`);
   }
+
+  return {
+    message: `El dispositivo ${deviceId} no está conectado o no se encuentra`,
+    status: 404,
+  };
 };
 
 exports.getActiveDevices = () => {
@@ -119,47 +152,42 @@ exports.getActiveDevices = () => {
 };
 
 exports.sendMessage = async (deviceId, numero, mensaje, imagen) => {
-  const connectionResult = await this.connectOrReconnect(deviceId);
+  const connectionResult = await this.connectOrReconnect(deviceId, true);
   const sock = socketManager.getSocket(deviceId);
 
-  if (connectionResult.status === 201) {
-    return res.status(404).json({
-      error: `El ID del dispositivo no se encuentra registrado`,
-      status: 404,
-    });
-  }
-
-  if (!sock) {
-    throw new Error(`El ID del dispositivo no se encuentra registrado`);
-  }
-
-  if (!numero) {
-    throw new Error(`El número de teléfono es requerido`);
-  }
-
-  if (!mensaje && !imagen) {
-    throw new Error(`El mensaje o la imagen es requerido`);
-  }
-
-  try {
-    const numeroWhatsApp = `${numero}@s.whatsapp.net`;
-
-    if (imagen) {
-      await sock.sendMessage(numeroWhatsApp, {
-        image: { url: imagen },
-        caption: mensaje || "",
-      });
-    } else {
-      await sock.sendMessage(numeroWhatsApp, { text: mensaje });
-    }
-
-    return { mensaje: "Mensaje enviado con éxito", status: 200 };
-  } catch (error) {
+  if (connectionResult.status === 200) {
     return {
-      mensaje: "Error al enviar el mensaje, El dispositivo no esta Vinculado",
+      error: `El ID del dispositivo no se encuentra registrado`,
       status: 404,
     };
   }
+
+  if (!numero) {
+    return {
+      error: `El número de teléfono es requerido`,
+      status: 404,
+    };
+  }
+
+  if (!mensaje && !imagen) {
+    return {
+      error: `El mensaje o la imagen es requerido`,
+      status: 404,
+    };
+  }
+
+  const numeroWhatsApp = `${numero}@s.whatsapp.net`;
+
+  if (imagen) {
+    await sock.sendMessage(numeroWhatsApp, {
+      image: { url: imagen },
+      caption: mensaje || "",
+    });
+  } else {
+    await sock.sendMessage(numeroWhatsApp, { text: mensaje });
+  }
+
+  return { mensaje: "Mensaje enviado con éxito", status: 200 };
 };
 
 exports.getUserInfo = async (deviceId) => {
